Fix: crash if player offer area exceeds 255 fields (possible on high resolutions)
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8823 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Arms Dealer Init.h
@@ -340,7 +340,7 @@ UINT32	GetArmsDealerItemTypeFromItemNumber( UINT16 usItem );
 //Count every single item the dealer has in stock
 //UINT32	CountTotalItemsInArmsDealersInventory( UINT8 ubArmsDealer );
 //Count only the # of "distinct" item types (for shopkeeper purposes)
-UINT16		CountTotalItemsRepairDealerHasInForRepairs( UINT8 ubArmsDealer );
+int			CountTotalItemsRepairDealerHasInForRepairs( UINT8 ubArmsDealer );
 
 void		AddObjectToArmsDealerInventory( UINT8 ubArmsDealer, OBJECTTYPE *pObject );
